requirements.txt
flask
faiss-cpu
sentence-transformers
PyPDF2
requests
